Cleanup .c/.cpp template programs to run all tests.
This commit standardizes the way all test_<...> programs are identified
and run. W/ no args, all test functions will be run. Apply this to
smart-pointers-CppCon-2019.cpp also.

Rework template programs to run all tests matching prefix.

This commit updates the loop in template programs that
look for a "test_<something>" name. With this fix, we now
support supplying a prefix test-case name, so that all
tests with that leading prefix substring will be run.