[Cpp-Coroutines-1] Fixes for compilation errors. With one patch.
This commit cleans-up code to address various compilation
errors. Now the sample program compiles, but seems to not fully
do what it was intended to do. The output looks like:

---
./coroutines-simplest-example: Hello World. (argc=1)
1. hello from foo1
1. hello from foo2
---

One patch applied to address this compilation error is to
comment out L71 below:

coroutines-simplest-example.cpp:71:17: error: 'this' argument to member function 'resume' has type 'const std::experimental::coroutine_handle<>', but function is not marked const
                corhdl.resume();
                ^~~~~~

---
 69         task_queue.push([start, corhdl, d = delay] {
 70             if (decltype(start)::clock::now() - start > d) {
 71                 corhdl.resume();    // <<<< Comment this out.
 72                 return true;
--

This allows the build to succeed, but this hack is fishy.